requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.2
djangorestframework==3.16.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
